arquivo

+++ questao1_itemb.py
@@ -1,12 +1,10 @@
 #encoding: UTF8
 from tkinter import *
 import time, os
-from threading import Thread, Semaphore, Lock
+from threading import Thread, Semaphore
 
 buffer1 = [""]                          # buffer1
 buffer2 = [""]                          # buffer2
-
-
 
 def interface(master,buffer1,buffer2,semaphore):                              # criacao das janelas
     fontePadrao = ("Arial", "10")                           # fonte padrao
@@ -43,18 +41,26 @@ def interface(master,buffer1,buffer2,semaphore):                              # 
     nomeLabel.config(wraplength=350)                        # quebra de linha
     nomeLabel.pack()
     time.sleep(.1)
-    # button = Button(master, text="Close", command=exit).pack()     # botao para fechar as janelas
 
     thread1 = Thread(target=refreshWrite)                   # thread de escrita
     thread2 = Thread(target=refreshRead)                    # thread de leitura
     thread1.start()
     thread2.start()
 
-root = Tk()
-w1 = Toplevel()                                             # cria multiplas janelas
+    if master is None:
+        os.popen("pkill -9 python")
+        os.waitpid()
 
-smf1 = Lock()
+root = Tk()
+w1 = Toplevel()
+root.title("Tela 1")# cria multiplas janelas
+w1.title("Tela 2")
+
+smf1 = Semaphore()
+
 interface(root, buffer1, buffer2, smf1)
 interface(w1, buffer2, buffer1, smf1)
+# print(root.state())
+
 root.mainloop()
-os.popen("pkill -9 python")
+os.popen("pkill python")
